Update PhonebookApp
Introduction
This document describes the implementation and functionality of the Phonebook Application developed using JavaFX. The program allows users to manage their contact list by adding, updating, deleting, searching, and sorting contacts, as well as displaying all contacts in a table view.
The application is designed with a simple Graphical User Interface (GUI) that lets users interact with the phonebook features. The contact data consists of a name and a phone number, and all information is displayed in a table for easy access.